# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User",
                "operationId": "a313b651590efd263cadf8cf2f6404e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "john@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "punt3n123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout User",
                "operationId": "9dce940f16581d4798e218faa1a8cd76",
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User",
                "operationId": "e9e4404b93a05e40dd3ee12377319a68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "username",
                                        "type": "string",
                                        "example": "john doe"
                                    },
                                    "first_name": {
                                        "description": "first name",
                                        "type": "string",
                                        "example": "john"
                                    },
                                    "last_name": {
                                        "description": "last name",
                                        "type": "string",
                                        "example": "doe"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "john@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "punt3n123"
                                    },
                                    "password_confirmation": {
                                        "description": "confirm password",
                                        "type": "string",
                                        "example": "punt3n123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/info": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Profile Info",
                "operationId": "b9c1de29e5caf14aadc5479df13f4f84",
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/upload-image": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload Image",
                "operationId": "66446f0ba46aef511107234ae26fe34a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/info": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "User Info",
                "operationId": "10f0af0a35d253175208440853dfe5f8",
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/update-password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update Password User",
                "operationId": "41a9db266913224df43fc738c92a9a6b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "current_password": {
                                        "description": "current password",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "new_password": {
                                        "description": "new password",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "new_password_confirmation": {
                                        "description": "new password",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Profile",
            "description": "Profile"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}